UI-92 ColorPicker/index.js
Extracts color methods into separate helpers.
