+++ src/Component/ColorPicker/Util/formatColor.js
@@ -1,0 +1,10 @@
+import hexFormat from './hexFormat';
+import rgbFormat from './rgbFormat';
+
+export default function formatColor(color, isRgbFormat) {
+    if (isRgbFormat) {
+        return rgbFormat(color);
+    }
+
+    return hexFormat(color);
+}
