+++ src/Component/ColorPicker/Util/isHex.js
@@ -1,0 +1,4 @@
+export default function isHex(color) {
+    return color.startsWith('#');
+}
+
