+++ src/Component/ColorPicker/Util/isRgb.js
@@ -1,0 +1,4 @@
+export default function isRgb(color) {
+    return color.startsWith('rgb');
+}
+
