+++ src/Component/ColorPicker/Util/rgbFormat.js
@@ -1,0 +1,11 @@
+import {colord} from 'colord';
+
+import isRgb from './isRgb';
+
+export default function rgbFormat(color) {
+    return isRgb(color)
+        ? color
+        : colord(color).toRgbString()
+    ;
+}
+
